requirements.txt
pyyaml
fastapi
uvicorn
mysqlclient
python-dotenv
sqlalchemy
pydantic